Added Material Light: a library of UI components (#223)
* Added Material Light: a library of UI components

There are several alternatives into the Material Design field for Angular 2 but this one has added value and advantages respect others in my opinion (Take a loot at the Readme.md for details). There is also an online demo to see it in action. (API documentation is still lacking but in process).

* Update README.md

I made a minor change in the description, but hope everything ok now.

Thanks to you and congratulations for the useful compilation of resources.

* Update README.md